Separate UI widgets from application-specific code
