+++ src/widget/number.js
@@ -87,7 +87,7 @@ function NumberUI(config) {
 	};
 
 	this.paint = function() {
-		if (value) span.value = value.toFixed(precision);
+		if (value != null) span.value = value.toFixed(precision);
 	};
 }
